started rewrite as singlepage app
emit does not work on mobile (when a new WebviewWindow is created the old one can no longer run code 😒). Therefore the new page (bookDetails) can not be told which book to load (I also tested writing it into the json directly and loading it in the new window but that also did not work).
Solution: Rewrite the app as a singlepage app with multiple screensized-divs (e.g. main or bookDetails) that overlap based on context.